Chris/connecting (#114)
* Minor fixed

* Pointers added

* All groups category pill padding

* Finished group admin frontend

* Edit buttons font size

* Most of the frontend for admin groups is done, but gonna switch to connecting more actions

* Removed imports from managetab

* Minor fixed

* Issue with signup/in pages

* Fixed component not showing

* Fixed signin/signup api, added all api routes to urls, fixed some links, grouppage now loads group

* Prop types for cur user

* Get data on pages

* More progress

* Finished group pages

* Fixed author data

* clearing search bar on groups page returns to prev tab

* Create group no longer has subscribers

* More prop types and stricter getuser

* Fixed height on thread comment

* Finished profile page

* Fixed comment card

* fixed follow/unfollow group

* fixed all groups page

* locked max screen size

* Started work on questions page

* ask me page done

+++ server/mongodb/actions/Group.js
@@ -1,17 +1,62 @@
 import mongoDB from "../index";
 import Group from "../models/Group";
+import User from "../models/User";
 
-export async function createGroup(name, description = "", tags = []) {
+export const followGroup = async (groupId, userId) => {
+  if (groupId == null || userId == null) {
+    throw new Error("All parameters must be provided!");
+  }
+
+  await mongoDB();
+
+  return User.findByIdAndUpdate(userId, { $push: { groups: groupId } });
+}
+
+export const unfollowGroup = async (groupId, userId) => {
+  if (groupId == null || userId == null) {
+    throw new Error("All parameters must be provided!");
+  }
+
+  await mongoDB();
+
+  return User.findByIdAndUpdate(userId, { $pull: { groups: groupId } });
+}
+
+export const createGroup = async (
+  currentUser,
+  name,
+  description,
+  category,
+  admin
+) => {
+  if (name == null || description == null || category == null || admin == null) {
+    throw new Error("All parameters must be provided!");
+  }
+
+  if (currentUser == null) {
+    throw new Error("You must be logged in to create a group!");
+  }
+
   await mongoDB();
 
   return Group.create({
     name,
     description,
-    tags,
+    category,
+    admin,
+  }).then(async group => {
+    const groupid = group._id;
+    await followGroup(groupid, currentUser);
+
+    return group;
   });
 }
 
-export async function deleteGroup(groupId) {
+export const deleteGroup = async (groupId) => {
+  if (groupId == null) {
+    throw new Error("All parameters must be provided!");
+  }
+
   await mongoDB();
 
   return Group.findOneAndDelete({ _id: groupId }).then(deletedGroup => {
@@ -20,32 +65,61 @@ export async function deleteGroup(groupId) {
     } else {
       return Promise.reject(new Error("No group matches the provided id"));
     }
+
     return deletedGroup;
   });
 }
 
-export async function followGroup(groupId, userId) {
+export const getGroup = async (groupId) => {
+  if (groupId == null) {
+    throw new Error("All parameters must be provided!");
+  }
+
   await mongoDB();
 
-  await Group.findByIdAndUpdate(groupId, { $push: { subscribers: userId } });
-}
-
-export async function unfollowGroup(groupId, userId) {
-  await mongoDB();
-
-  await Group.findByIdAndUpdate(groupId, { $pull: { subscribers: userId } });
-}
-
-export const getGroup = groupId =>
-  Group.findById(groupId).then(group => {
+  return Group.findById(groupId).then(async group => {
     if (group == null) {
       throw new Error("Group does not exist!");
     }
+
+    const people = await User.find({ groups: groupId }).count();
 
     return {
       id: group._id,
       name: group.name,
       description: group.description,
+      admin: group.admin,
+      subscribers: group.subscribers,
       image: group.image,
+      category: group.category,
+      people,
     };
   });
+}
+
+export const searchGroups = async (groupId) => {
+  if (groupId == null) {
+    throw new Error("All parameters must be provided!");
+  }
+
+  await mongoDB();
+
+  return Group.findById(groupId).then(async group => {
+    if (group == null) {
+      throw new Error("Group does not exist!");
+    }
+
+    const people = await User.find({ groups: groupId }).count();
+
+    return {
+      id: group._id,
+      name: group.name,
+      description: group.description,
+      admin: group.admin,
+      subscribers: group.subscribers,
+      image: group.image,
+      category: group.category,
+      people,
+    };
+  });
+}
